put initial fetch in it's own file
